Рефакторинг. Утилитарные функции перенесены в utils/, разбиты на смысловые файлы
